+++ src/commonTest/kotlin/csense/kotlin/extensions/collections/generic/ForEachTest.kt
@@ -1,0 +1,4 @@
+package csense.kotlin.extensions.collections.generic
+
+class ForEachTest {
+}