+++ src/commonTest/kotlin/csense/kotlin/extensions/collections/set/SetTest.kt
@@ -1,0 +1,4 @@
+package csense.kotlin.extensions.collections.set
+
+class SetTest {
+}