+++ src/commonTest/kotlin/csense/kotlin/extensions/generic/foreachTest.kt
@@ -1,0 +1,4 @@
+package csense.kotlin.extensions.generic
+
+class foreachTest {
+}