+++ src/jvmTest/kotlin/csense/kotlin/extensions/WeakReferenceTest.kt
@@ -1,0 +1,4 @@
+package csense.kotlin.extensions
+
+class WeakReferenceTest {
+}